feat: refactor frontend
