+++ frontend/src/components/Editor/types.txt
