+++ frontend/src/utils.ts
@@ -1,4 +1,5 @@
 import lzString from "lz-string";
+import { Examples } from "./constants";
 
 export const decodeCode = (code: string | null): string => {
   if (!code) {
@@ -43,5 +44,27 @@ export const trackEvent = (): void => {
     window.gtag('event', "execute", {
       'event_category': "engagement",
     });
+  }
+}
+
+export const determineCode = async (setCode: ((code: string) => void)): Promise<void> => {
+  const urlParams = new URLSearchParams(window.location.search);
+  const localStorageCode = window.localStorage.getItem("code")
+  // TODO: remove (if: code) after a couple of months. Was kept for backwards compatibility
+  if (urlParams.has("code")) {
+    const newCode = decodeCode(urlParams.get("code"))
+    setCode(newCode)
+  } else if (urlParams.has("s")) {
+    const key = urlParams.get("s")
+    if (key) {
+      const sharedCode = await fetchSharedCode(key)
+      if (sharedCode) {
+        setCode(sharedCode)
+      }
+    }
+  } else if (localStorageCode) {
+    setCode(localStorageCode)
+  } else {
+    setCode(Examples[0].code)
   }
 }
